implement proofshare bmgr do
